Extract display session and image pipeline from bl.cpp
bl.cpp mixed boot, Wi-Fi, API refresh, image I/O, sleep, and X gestures
(~3.8k lines), which blocked review and follow-on product work.

Move the server refresh path into dedicated modules:
- display_session: loadApiDisplayInputs, downloadAndShow,
  handleApiDisplayResponse, https error name table
- image_pipeline: stream download, SPIFFS write, NVS path/playlist
  helpers, previous-frame EPD load, cache existence check

bl.cpp retains orchestration, captive portal, setup credentials, sleep,
and touchbar. Shared session state is documented in bl_bridge.h for
later shrink.

Also: fix previous-image EPD load (was a no-op), consolidate path
persistence, add #pragma once on api-client/display.h, declare
png_to_epd in display.h, and document the map/next steps in
docs/architecture-modules.md and the README.
